sas_client/requirements.txt
pyyaml
requests
urllib3
ntnx-clustermgmt-py-client==4.0.1
ntnx-networking-py-client==4.0.1
ntnx-prism-py-client==4.0.1
ntnx-vmm-py-client==4.0.1